+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import kurdish_localization_plus
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FlutterKurdishLocalizationPlugin.register(with: registry.registrar(forPlugin: "FlutterKurdishLocalizationPlugin"))
+}